+++ MmfReader.cs
@@ -1,11 +1,9 @@
 ﻿using System;
 using System.IO;
 using System.IO.MemoryMappedFiles;
-using System.Linq.Expressions;
 using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
-using SimHub.Plugins;
 
 namespace F1Manager2024Plugin
 {